+++ CHANGELOG.md
@@ -43,6 +43,12 @@ For more information on SemVer, please visit http://semver.org/.
 * Added `WPDKMath::isInfinity()` public static method to replace `wpdk_is_infinity()` the inline (deprecated) function
 * Added `WPDKHTMLTagSelect::selected()` public static method to replace `wpdk_selected()` the inline (deprecated) function
 * Added `label` key for `WPDKUIControlSubmit` control
+* Added new `WPDKPreferences`, `WPDKPreferencesBranch`, `WPDKPreferencesViewController` and `WPDKPreferencesView`
+
+#### Deprecated
+
+* `WPDKConfiguration` use new `WPDKPreferences`
+* `WPDKConfigurationView` use new `WPDKPreferencesViewController` and `WPDKPreferencesView`
 
 
 ## Version 1.1.2